feat: rewrite update script
Rewriting in nushell makes it much easier to interact with structured
data like API responses, and allows easy/simpler implementation of
pagination, filtering, etc. we need. I think this will also be a little
faster because nix-prefetch is called in parallel for all archs.

Also implemented pagination in this PR which means we should always find
a release. Might fix #50, but the comments there are a bit confusing.

Also fixes #60.